feat: add search unmount
