+++ vite-project/package.json
@@ -8,11 +8,11 @@
     "build": "tsc && vite build",
     "preview": "vite preview",
     "pretest": "eslint .",
-    "test": "react-scripts test",
     "coverage": "react-scripts test --coverage"
   },
   "dependencies": {
     "@testing-library/user-event": "^14.4.3",
+    "eslint-config-prettier": "^8.8.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.9.0",
@@ -28,6 +28,7 @@
     "@typescript-eslint/parser": "^5.55.0",
     "@vitejs/plugin-react": "^3.1.0",
     "eslint": "^8.36.0",
+    "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-react": "^7.32.2",
     "eslint-plugin-react-hooks": "^4.6.0",
     "husky": "^8.0.3",
