+++ vite-project/prettierrc.json
@@ -1,7 +1,7 @@
 {
     "endOfLine": "auto",
     "semi": true,
-    "singleQuote": false,
+    "singleQuote": true,
     "tabWidth": 2,
     "trailingComma": "es5",
     "printWidth": 100,
