+++ vite-project/src/componets/css/App.css
@@ -37,11 +37,13 @@
 }
 h1 {
   color: rgb(214, 214, 75);
+  margin-bottom: 0;
 }
 h2 {
   color: rgb(85, 85, 223);
 }
 .ATM__grid {
+    margin-top: 20px;
     border: 2px solid yellow;
     border-radius: 17px;
     display: grid;
@@ -84,6 +86,7 @@ h2 {
   .search {
     height: 27px;
     border: 1px solid yellow;
+    border-radius: 10px;
   }
   .search__magnifier {
     width: 20px;
