--- vite-project/src/componets/shop.tsx
@@ -1,29 +1,0 @@
-function Atmshop() {
-  return (
-    <div>
-      <h1 className="shop__name">Bitcoin ATM shop</h1>
-      <div className='conteiner'>
-        <div className="ATM__grid">
-          <div>
-            <p className="class__description"><h2>Genesis-1</h2><img src="../src/assets/genesis1.webp" className="ATM__first" alt="first ATM" /> The latest model from Lamassu, designed for simplicity and elegance with its steel construction. </p>
-          </div>
-          <div >
-            <p className="class__description"><h2>Tejo</h2><img src="../src/assets/2sintra-front.png" className="ATM__second" alt="first ATM" /> The latest model from Lamassu, designed for simplicity and elegance with its steel construction. </p>
-          </div>
-          <div>
-            <p className="class__description"><h2>Satoshi-1</h2><img src="../src/assets/satoshi1.webp" className="ATM__third" alt="first ATM" /> The latest model from Lamassu, designed for simplicity and elegance with its steel construction. </p>
-          </div>
-        </div>
-      </div>
-
-      <div className="card">
-
-
-      </div>
-
-
-    </div>
-  );
-}
-
-export default Atmshop
